New functionality to create a new job and save user

+++ backend/job/views.py
@@ -1,10 +1,12 @@
 from webbrowser import get
 from django.shortcuts import render
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 from django.db.models import Avg, Min, Max, Count
 from rest_framework.pagination import PageNumberPagination
+
+from rest_framework.permissions import IsAuthenticated
 
 from .serializers import JobSerializer
 from .models import Job
@@ -45,7 +47,9 @@ def getJob(request, pk):
 
 # Create a new job
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
     
     job = Job.objects.create(**data)
@@ -55,8 +59,12 @@ def newJob(request):
 
 # Update Job
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    
+    if job.user != request.user:
+        return Response({ 'message': 'You Cannot Update This Job!' }, status=status.HTTP_403_FORBIDDEN)
     
     job.title = request.data['title']
     job.description = request.data['description']
@@ -78,8 +86,12 @@ def updateJob(request, pk):
 
 # Delete a job
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    
+    if job.user != request.user:
+        return Response({ 'message': 'You Cannot Delete This Job!' }, status=status.HTTP_403_FORBIDDEN)    
     
     job.delete()
